write authorization logic

+++ app/Http/Controllers/StudentController.php
@@ -3,15 +3,29 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreStudentRequest;
+use App\Role;
 use App\User;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use Illuminate\Session\Store;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class StudentController extends Controller
 {
+    private $studentRoleId;
+
+    /**
+     * StudentController constructor.
+     */
+    public function __construct()
+    {
+        $this->studentRoleId = Role::whereName('Student')->first()->id;
+
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -19,7 +33,11 @@ class StudentController extends Controller
      */
     public function index()
     {
-        return User::all()->toJson();
+        if (Gate::denies('all-student-info')) {
+            return User::where('role_id', $this->studentRoleId)->where('id', Auth::user()->id)->get()->toJson();
+        } else {
+            return User::where('role_id', $this->studentRoleId)->get()->toJson();
+        }
     }
 
     /**
@@ -40,13 +58,16 @@ class StudentController extends Controller
      */
     public function store(StoreStudentRequest $request)
     {
+        if (Gate::denies('store-student')) {
+            abort(403);
+        }
         $data = $request->all();
         $data['password'] = bcrypt($request->get('idn'));
-        //dd($data);
-        $student = new User;
+        $data['role_id'] = $this->studentRoleId;
 
+        $student = new User;
         $student->create($data);
-        //dd($student);
+
         return $student->toJson();
     }
 
@@ -58,8 +79,11 @@ class StudentController extends Controller
      */
     public function show($id)
     {
+        if (Gate::denies('see-own-student-info',$id)) {
+            abort(403);
+        }
         $student = User::findOrFail($id);
-        $data = ['student'=>$student];
+        $data = ['student' => $student];
         return view('show_student')->with($data);
     }
 
@@ -83,6 +107,9 @@ class StudentController extends Controller
      */
     public function update(StoreStudentRequest $request, $id)
     {
+        if (Gate::denies('edit-student')) {
+            abort(403);
+        }
         $student = User::findOrFail($id);
         $student->update($request->all());
         return $student;
@@ -96,6 +123,9 @@ class StudentController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::denies('remove-student')) {
+            abort(403);
+        }
         User::destroy($id);
     }
 }
